+++ apps/server/collaboration/src/main/java/com/guanxian/platform/collaboration/CollaborationStore.java
@@ -8,7 +8,7 @@ import java.util.UUID;
 
 interface CollaborationStore {
     List<CollaborationView> list(
-            ActorScope actor, String query, boolean includeDeleted, int offset, int limit);
+            ActorScope actor, String query, boolean includeDeleted, long offset, int limit);
 
     long count(ActorScope actor, String query, boolean includeDeleted);
